+++ internal/agent/f5/as3/proxy_protocol_v2.go
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: Copyright 2025 SAP SE or an SAP affiliate company
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package as3
+
+import _ "embed"
+
+//go:embed proxy_protocol_v2.irule
+var pp2 string